+++ src/rpclib/service.py
@@ -127,8 +127,15 @@ class ServiceBase(object):
         return cls.__name__
 
     @classmethod
+    def get_service_key(cls):
+        return '{%s}%s' % (cls.get_tns(), cls.get_service_name())
+
+    @classmethod
     def get_service_name(cls):
-        return cls.__service_name__
+        if cls.__service_name__ is None:
+            return cls.__name__
+        else:
+            return cls.__service_name__
 
     @classmethod
     def get_port_types(cls):